--- src/repository/index.js
@@ -1,5 +1,0 @@
-const ContactsRepository = require('./contacts');
-
-module.exports = {
-  ContactsRepository,
-};
