--- src/services/index.js
@@ -1,4 +1,0 @@
-const ContactsService = require('./contacts');
-module.exports = {
-  ContactsService,
-};
